llama3_finetune/requirements.txt
pytorch-lightning==2.1.0
optimum-neuron==0.0.21
regex
tensorboard
datasets
sentencepiece
nltk
neuronx_distributed
ray[data,train,tune,serve]